+++ src/main/java/com/zerobase/foodlier/common/redis/service/RefreshTokenService.java
@@ -1,13 +1,35 @@
 package com.zerobase.foodlier.common.redis.service;
 
+import com.zerobase.foodlier.common.redis.domain.model.RefreshToken;
 import com.zerobase.foodlier.common.redis.dto.RefreshTokenDto;
+import com.zerobase.foodlier.common.redis.repository.RefreshTokenRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface RefreshTokenService {
 
-    boolean isRefreshTokenExisted(String email);
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class RefreshTokenService {
 
-    void save(RefreshTokenDto refreshTokenDto);
+    private final RefreshTokenRepository refreshTokenRepository;
 
-    void delete(String email);
+    public boolean isRefreshTokenExisted(String email) {
+        return refreshTokenRepository.existsById(email);
+    }
 
+    public void save(RefreshTokenDto refreshTokenDto) {
+        refreshTokenRepository.save(RefreshToken.builder()
+                .email(refreshTokenDto.getMemberEmail())
+                .refreshToken(refreshTokenDto.getRefreshToken())
+                .expiration(refreshTokenDto.getTimeToLive())
+                .build());
+    }
+
+    public void delete(String email) {
+        if (refreshTokenRepository.existsById(email)) {
+            refreshTokenRepository.deleteById(email);
+        }
+    }
 }
