+++ src/main/java/com/zerobase/foodlier/common/redisson/service/RedissonLockService.java
@@ -1,9 +1,46 @@
 package com.zerobase.foodlier.common.redisson.service;
 
-public interface RedissonLockService {
+import com.zerobase.foodlier.common.redisson.exception.RedissonException;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
+import org.springframework.stereotype.Service;
 
-    void lock(String group, String key);
+import java.util.concurrent.TimeUnit;
 
+import static com.zerobase.foodlier.common.redisson.exception.RedissonErrorCode.LOCK_ERROR;
 
-    void unlock(String group, String key);
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class RedissonLockService {
+    private final RedissonClient redissonClient;
+
+    private final long WAIT_TIME = 3L;
+    private final long LEASE_TIME = 1L;
+    private final TimeUnit TIME_UNIT = TimeUnit.SECONDS;
+
+    public void lock(String group, String key) {
+        RLock lock = redissonClient.getLock(getLockKey(group, key));
+
+        try {
+            boolean isLock = lock.tryLock(WAIT_TIME, LEASE_TIME, TIME_UNIT);
+            if (!isLock) {
+                throw new RedissonException(LOCK_ERROR);
+            }
+        } catch (RedissonException e) {
+            throw e;
+        } catch (Exception e) {
+            log.error("Redisson lock failed");
+        }
+    }
+
+    public void unlock(String group, String key) {
+        redissonClient.getLock(getLockKey(group, key)).unlock();
+    }
+
+    private static String getLockKey(String group, String key) {
+        return "Lock : " + group + "-" + key;
+    }
 }
